pyproject.toml
[project]
name = "fingersalesai"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-multipart",
    "python-dotenv",
    "pydantic",
    "langgraph",
    "langchain-core",
    "langchain-google-genai",
    "langchain-community",
    "tavily-python",
    "mysql-connector-python",
    "bcrypt",
]
